fix point structure, add distance calculation to worker state, move SystemState and WorkerState to a separate package
