+++ geo-monitoring-api/src/main/java/com/adg/geomonitoringapi/event/controller/ReportController.java
@@ -1,6 +1,6 @@
 package com.adg.geomonitoringapi.event.controller;
 
-import com.adg.geomonitoringapi.event.SystemState;
+import com.adg.geomonitoringapi.state.SystemState;
 import com.adg.geomonitoringapi.event.repository.EventRepository;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.web.bind.annotation.GetMapping;
