+++ geo-monitoring-api/src/main/java/com/adg/geomonitoringapi/event/entity/TaskAssignedEvent.java
@@ -3,7 +3,7 @@ package com.adg.geomonitoringapi.event.entity;
 import com.adg.geomonitoringapi.event.Task;
 import com.adg.geomonitoringapi.event.TaskStatus;
 import com.adg.geomonitoringapi.worker.entity.Worker;
-import com.adg.geomonitoringapi.event.SystemState;
+import com.adg.geomonitoringapi.state.SystemState;
 import jakarta.persistence.*;
 import lombok.*;
