+++ geo-monitoring-api/src/main/java/com/adg/geomonitoringapi/event/entity/TaskCreatedEvent.java
@@ -2,7 +2,7 @@ package com.adg.geomonitoringapi.event.entity;
 
 import com.adg.geomonitoringapi.event.Task;
 import com.adg.geomonitoringapi.event.TaskStatus;
-import com.adg.geomonitoringapi.event.SystemState;
+import com.adg.geomonitoringapi.state.SystemState;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
 import lombok.*;
